Moved `space_view_button` closer to where it's still used
