+++ crates/re_viewport/src/lib.rs
@@ -22,33 +22,3 @@ pub use viewport_blueprint::ViewportBlueprint;
 pub mod external {
     pub use re_space_view;
 }
-
-// ---------------------------------------------------------------------------
-
-// TODO(andreas): This should be part of re_data_ui::item_ui.
-pub mod item_ui {
-    use re_data_ui::item_ui;
-    use re_viewer_context::{Item, ViewerContext};
-
-    use crate::space_view::SpaceViewBlueprint;
-
-    pub fn space_view_button(
-        ctx: &mut ViewerContext<'_>,
-        ui: &mut egui::Ui,
-        space_view: &SpaceViewBlueprint,
-    ) -> egui::Response {
-        let item = Item::SpaceView(space_view.id);
-        let is_selected = ctx.selection().contains(&item);
-
-        let response = ctx
-            .re_ui
-            .selectable_label_with_icon(
-                ui,
-                space_view.class(ctx.space_view_class_registry).icon(),
-                space_view.display_name.clone(),
-                is_selected,
-            )
-            .on_hover_text("Space View");
-        item_ui::cursor_interact_with_selectable(ctx, response, item)
-    }
-}
